Zip Function Filepath
Eric,  When I first installed and ran the tool, the zip filepath wasn't set on my machine which led to issues.  It may have been because of the way I set up my R, but if this happens to anyone else, I've included a line of code into the SPF-R under line 4 that will allow the user to reset the zip function filepath to the proper etting to allow for the program to run.  This was the issue I have been telling you about that I hadn't commented on when I told you I would a while back, so I apologize.